// src/features/dashboard/types/dashboard.types.ts
export type LocalizedField = string | LocalizedName;
// ...
export interface OrderProduct {
  id: number;
  name: LocalizedField;
  pivot: { product_quantity: number };
}

export interface OrderUser {
  id: number;
  name: LocalizedField;
  email: string;
}

export interface RecentOrder {
  id: number;
  name: string;
  status: string;
  total_price: number;
  created_at: string;
  products: OrderProduct[];
  user: OrderUser;
}